test(snapshot): convert tests to typescript
