--- tests/unit/extensions/snapshot/package.json
@@ -1,3 +1,0 @@
-{
-    "type": "commonjs"
-}
